infra/requirements.txt
# AWS CDK — Python
aws-cdk-lib>=2.150.0
constructs>=10.0.0
